use pangolinmqtt

+++ CODE/shapesp/i2c.ino
@@ -44,7 +44,7 @@ uint8_t i2c_read_uint8(uint8_t address, uint8_t reg) {
 };
 
 uint8_t i2c_read_uint8(uint8_t address, uint16_t reg) {
-    uint8_t buffer[2] = {((reg>>8)&0xFF),reg&0xFF};
+    uint8_t buffer[2] = {(uint8_t)((reg>>8)&0xFF),(uint8_t)(reg&0xFF)};
     brzo_i2c_start_transaction(address, _i2c_scl_frequency);
     brzo_i2c_write(buffer, 2, false);
     brzo_i2c_read(buffer, 1, false);
